edit-upload

+++ resources/views/manage.blade.php
@@ -3,19 +3,25 @@
 @section('title','Praktikum Web')
 <div class="menuManage">
     <tr>
+        <td><a href="/">Home</a></td>
         <td><a href="/manage">Manage Artikel</a></td>
         <td><a href="/manage/user">Manage User</a></td>
         <td><a href="/pages" target="_blank">Cetak PDF</a></td>
-    </tr>    
+    </tr>
 </div>
+
 @section('judul','Manage Artikel')
 
 @section('post')
 
-<div class="tambah mb-2">
-    <a href="/add" class="btn btn-primary">Tambah Artikel</a>
-</div>
-
+    <div class="tambah mb-2">
+        <a href="/add" class="btn btn-primary">Tambah Artikel</a>
+    </div>
+    @if(session('status'))
+    <div class="alert alert-success">
+        {{session('status')}}
+    </div>
+@endif
 <table class="table table-bordered table-striped ">
     <thead class="thead-dark">
         <tr>
@@ -32,7 +38,7 @@
             <td>{{$a->title}}</td>
             <td>{{$a->created_at}}</td>
             <td><a href="/edit/{{ $a->id }}" class="badge badge-warning">Edit</a>
-                <a href="article/delete/{{ $a->id }}" class="badge badge-danger">Delete</a>
+                <a href="article/delete/{{ $a->id }}" class="badge badge-danger" onclick="return confirm('Yakin Ingin menghapus Artikel???')">Delete</a>
             </td>
         </tr>
         @endforeach
